Se agregaron el db context todos los elementos pertinente a las tablas de las base de datos, se realizaron los cambios pertinentes en el mecanismo de navegacion por parte de entity framework core realizando asi los join de las proyeciones que son requeridas en cada uno de los metoods de consulta de data, tambien se elliminaron elementos no necesarios para el aplicativo que se encuentra en desarrollo que es el de escritorio.

+++ Core/AeroVeloz.Domain/Entities/Users/Permission/Permission.cs
@@ -2,7 +2,7 @@
 
 namespace AeroVeloz.Domain.Entities.Users.Permission
 {
-    public partial class Permission : BEntity<int>
+    public partial class Permission : BEntity<byte>
     {
         public string? codePermision { get; init; }
         public string? description { get; init; }
